Fix gh36481: Replace HandlerMappingIntrospector with PreFlightRequestHandler bean in MVC config
